pyproject.toml
[tool.poetry]
name = "disk0muzik"
version = "0.1.0"
description = ""
authors = ["Jared Robertson <robertsonjareds@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
spotipy = "^2.24.0"
yt-dlp = "^2024.7.9"
python-dotenv = "^1.0.1"
pynacl = "^1.5.0"
colorlog = "^6.8.2"
aiofiles = "^24.1.0"
psycopg2 = "^2.9.9"
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
